// models/Filter.ts
export interface Filter {
    city: string;
    cityValue: string;
    country: string;
    countryValue: string;
    name: string;
    seniority: string;
    seniorityValue: string;
    project: string;
    projectValue: string;
    projectManager: string;
    pmValue: string;
    technology: string;
    technologyValue: string;
}